Surface reset fixes (#27)
* Linter claims log tag can be at most 23 chars only

* Don't blindly add the target view on surface creation

Indeed, we might be in a situation where the surface has been destroyed
and created again. In such a case the target view was already added to
the layout. If we detect the situation, we simply don't add the target
view again.

* Rework FrameAnalyserManager handling

We now make sure it's properly resurrected after a cleanUp() call (e.g.
when the surface has been destroyed). Also make sure we didn't forget
about the symbologies which were previously requested, we now add them
again when the manager is recreated.