Addon providers (#28)
* Use services to load license encumbered providers

* Move the license encumbered sdk packages in their own modules

* Disable license encumbered modules if their sdk isn't provided

We simply find out if there's any .jar or .aar in the libs directory of
the modules, if we don't find any then the corresponsing module tasks
are all disabled and the demo app doesn't bundle said module. This way
we ensure a proper build out of the box and extended features as more of
the proprietary SDKs are provided.

* Remove the license encumbered SDKs since they are optional

Also add potential matches for the SDK files to the git ignore list.