Remove device specific hacks for CV1 #84
No crash or performance issue was observed from removing the hack, so
I assume their original purpose was to stop the adaptive resolution from
going all the way to the lowest resolution (as no performance
improvement happens each step). To solve this, a minimum resolution was
added as the preview would become unusable below that point anyway.