#84 - Remove device-specific camera hacks (#121)
- Remove device-specific hacks in CameraV1.

As no crash or performance reduction was observed when removing the hacks, I assume the original purpose was to stop the adaptive resolution from being called repeatedly until the minimum resolution was reached, as lower resolutions do not seem to improve performance on these devices. To solve this, a minimum resolution is now enforced (720px minimum on the Y axis) below which the preview is considered unusable anyway (though 480px would likely still be fine). This means even if the adaptive resolution is called over and over, it will stop at a much more reasonable point.

Tested with SPA43LTE, no access to LG-H340n or Archos Sense 50X currently.

- No device-specific hacks were found in CameraV2.